# Optimize elasticity evaluation in material model