--- a/source/simulator/simulator_access.cc
+++ b/source/simulator/simulator_access.cc
@@ -373,12 +373,7 @@ namespace aspect
   const FiniteElement<dim> &
   SimulatorAccess<dim>::get_fe () const
   {
-    Assert (simulator->dof_handler.n_dofs() > 0,
-            ExcMessage("You are trying to access the FiniteElement before the DOFs have been "
-                       "initialized. This may happen when accessing the Simulator from a plugin "
-                       "that gets executed early in some cases (like material models) or from "
-                       "an early point in the core code."));
-    return simulator->dof_handler.get_fe();
+    return simulator->finite_element;
   }