@gassmoeller 2021-11-23
I had to change the initialization of FEPointEvaluation, because in models with mesh deformation the mapping is changed after the material model is initialized. But now everything works, ready from my side.